refactor(login/comp): refac
